Parser

The Parser Monad

data Parser a

A parser is simply a wrapper around a function transforming input to a | list of valid Parses.

Running the parser

parse :: Parser a -> String -> Maybe a

parseStart :: Parser a -> String -> Maybe (String, a)

parsePartial :: Parser a -> String -> SingleResult a

Getting a list of possible parses

listParse :: Parser a -> String -> [a]

listParseStart :: Parser a -> String -> [(String, a)]

listParsePartial :: Parser a -> String -> ListResult a

Parsing bytes

item :: Parser Char

Match any character.

end :: Parser ()

Match the end of the input. Note that this also matches at the end of the current input block when partial parsing is enabled, if further parsers are chained, the result is Incomplete, instead of Failed.

char :: Char -> Parser Char

Parse exactly the given character.

satisfy :: (Char -> Bool) -> Parser Char

Parse a char matching the predicate.

oneOf :: [Char] -> Parser Char

Parse any of the given characters.

noneOf :: [Char] -> Parser Char

Parse any character other than one given.

Lookahead

ahead :: Parser a -> Parser a

Run a parser, then reset the input (lookahead).

peek :: Parser Char

Match any character with lookahead.

Parsing sequences

sepBy :: Alternative f => f a -> f sep -> f [a]

Parse a list of items by px, separated with the separator psep. If no items can be found, the empty list is returned.

sepBy1 :: Alternative f => f a -> f sep -> f [a]

Parse a list of items by px, separated with the separator psep. If no items can be found, the parser fails.

chainl :: (Alternative m, Monad m) => m a -> m (b -> a -> b) -> b -> m b

chainl p op x parses zero or more occurrences of p, separated by op. Returns a value produced by a left associative application of all functions returned by op, starting with the value x. If there are no occurrences of p, x is returned.

chainl1 :: (Monad m, Alternative m) => m b -> m (b -> b -> b) -> m b

chainl1 p op parses one or more occurrences of p, separated by op. Returns a value produced by a left associative application of all functions returned by op.

(<:>) :: Applicative f => f a -> f [a] -> f [a]

Cons a parsed value onto a parsed list.

string :: String -> Parser String

Parse an exact string.