Parser.BaseParsers

Documentation

item :: Parser Char

Match any character.

ahead :: Parser a -> Parser a

Run a parser, then reset the input (lookahead).

peek :: Parser Char

Match any character with lookahead.

end :: Parser ()

Match the end of the input. Note that this also matches at the end of the current input block when partial parsing is enabled, if further parsers are chained, the result is Incomplete, instead of Failed.