Parser.Combinators

Parsing bytes

char :: Char -> Parser Char

Parse exactly the given character.

satisfy :: (Char -> Bool) -> Parser Char

Parse a char matching the predicate.

oneOf :: [Char] -> Parser Char

Parse any of the given characters.

noneOf :: [Char] -> Parser Char

Parse any character other than one given.

Parsing sequences

sepBy :: Alternative f => f a -> f sep -> f [a]

Parse a list of items by px, separated with the separator psep. If no items can be found, the empty list is returned.

sepBy1 :: Alternative f => f a -> f sep -> f [a]

Parse a list of items by px, separated with the separator psep. If no items can be found, the parser fails.

chainl :: (Alternative m, Monad m) => m a -> m (b -> a -> b) -> b -> m b

chainl p op x parses zero or more occurrences of p, separated by op. Returns a value produced by a left associative application of all functions returned by op, starting with the value x. If there are no occurrences of p, x is returned.

chainl1 :: (Monad m, Alternative m) => m b -> m (b -> b -> b) -> m b

chainl1 p op parses one or more occurrences of p, separated by op. Returns a value produced by a left associative application of all functions returned by op.

(<:>) :: Applicative f => f a -> f [a] -> f [a]

Cons a parsed value onto a parsed list.

string :: String -> Parser String

Parse an exact string.