Parser.Result

Documentation

data SingleResult a

The result of a possibly partial parse, where only the first successful parse is given, if present.

Constructors

SIncomplete
The result is not a valid parse, but it is a prefix of a valid parse.

SParsed a
The parse was successful, and the first successful parse is given.

SFailed
The result is not a valid parse.

data ListResult a

The result of a possibly partial parse, where every successful parse is given, if present.

Constructors

LIncomplete
The result is not a valid parse, but it is a prefix of a valid parse.

LParsed [a]
The parse was successful, and every successful parse is given in a list.

LFailed
The result is not a valid parse.

class Results r

A typeclass for all the ways to receive the result of a partial parse.

Minimal complete definition

results

results :: Results r => Parses a -> r a

Produce a result of a partial parse.