Parser.Types

Documentation

data Termination

Termination refers to whether the input is known to terminate, or could | be continued, e.g. from a REPL.

Constructors

Closed

Open

type Input = (Termination, String)

Input is passed around internally as a tuple of the termination type and | and the flag.

type Parses a = [Maybe (String, a)]

The result of a parse is internally stored as a list of optional tuples | of the remaining string and the parsed object. A value of Nothing within | the list refers to a partial parse: the string is the prefix of a valid | parse. This is, however, not returned to the user unless there are no | valid complete parses.

newtype Parser a

A parser is simply a wrapper around a function transforming input to a | list of valid Parses.

Constructors

Parser

Fields

runParser :: Input -> Parses a